Some unknown compicated fixes.

+++ app/Http/Controllers/Api/Admin/CarController.php
@@ -2,19 +2,19 @@
 
 namespace App\Http\Controllers\Api\Admin;
 
+use App\Entity\Car;
 use App\Manager\CarManager;
 use Illuminate\Http\Request;
-use App\Http\Controllers;
+use App\Http\Controllers\Controller;
 
 
-class CarController extends Controllers\Controller
+class CarController extends Controller
 {
     protected $carManager;
 
     public function __construct(CarManager $carManager)
     {
         $this->carManager = $carManager;
-        $this->middleware('auth');
     }
 
     /**
@@ -22,6 +22,8 @@ class CarController extends Controllers\Controller
      * Method: GET
      */
     public function index() {
+        $this->authorize('create', new Car());
+
         return $this->getCarList();
     }
 
@@ -30,6 +32,9 @@ class CarController extends Controllers\Controller
      * Method: POST
      */
     public function store(Request $request) {
+        $this->authorize('create', new Car());
+
+
         $res = $this->carManager->saveCar($request);
 
         return response()->json($res->toArray());
@@ -40,6 +45,9 @@ class CarController extends Controllers\Controller
      * Method: GET
      */
     public function show($id) {
+        $this->authorize('create', new Car());
+
+
         return $this->getCar($id);
     }
 
@@ -48,6 +56,8 @@ class CarController extends Controllers\Controller
      * Method: PUT/PATCH
      */
     public function update(Request $request, int $id) {
+        $this->authorize('create', new Car());
+
         $oCar = $this->cars->getById($id);
         $res = $this->carManager->saveCar($request);
 
@@ -59,6 +69,8 @@ class CarController extends Controllers\Controller
      * Method: DELETE
      */
     public function destroy(int $id) {
+        $this->authorize('create', new Car());
+
         $res = $this->carManager->deleteCar($id);
         return response("", 200);
     }
